refactor : menuRequestDto 네이밍 및 필드 수정
세부사항
- menuRequestDto 내 id 필드 및 booth 필드 제거
- menu 내 price 필드 Long에서 int로 변경

+++ src/main/java/likelion/festival/menu/domain/Menu.java
@@ -28,20 +28,20 @@ public class Menu {
 
     private String name;
 
-    private Long price;
+    private int price;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "booth_id", nullable = false, foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT))
     private Booth booth;
 
-    public Menu(final Long id, final String name, final Long price, final Booth booth) {
+    public Menu(final Long id, final String name, final int price, final Booth booth) {
         this.id = id;
         this.name = name;
         this.price = price;
         this.booth = booth;
     }
 
-    public static Menu forSave(final String name, final Long price, final Booth booth) {
+    public static Menu forSave(final String name, final int price, final Booth booth) {
         return new Menu(null, name, price, booth);
     }
 }
